Fix selected for User#commit_email input
This fixes the input for User#commit_email so the correct option is
selected. The problem was that if no database value was set, a default
value was set as the selected option. We don't want to set a database
value without user input, so this removes the default value.
User sees clear information that their primary email will be used
as their default commit email.

Changelog: fixed