Update upcoming reconciliation banner copy
We want to be able to flexibly change the dates/timing around the
reconciliation notifications while we figure out how this should work.
Update the copy to instruct the user to view the docs for this
information so that we can more easily keep it up to date.

Issue: https://gitlab.com/gitlab-org/customers-gitlab-com/-/issues/3474