Only include visible groups a project is shared with
This redacts the list of groups a project is shared with, removing
groups the current user does not have read access to.

The redaction process introduces a N+1 regression due to the need to
evaluate `Group.root_ancestor.saml_provider`. Methods are added to
address this manually, since this cannot be addressed with an AR
preload, as we do in the `API::Entities::Project` class, which otherwise
manages preloading for entities.

Changelog: security