Only report MysteryRedisDurationError for shared state Redis
This error reporting (behind a feature flag) was added to investigate
high durations we saw on the shared state Redis on GitLab.com. However,
it did not check the Redis instance. This is a particular problem with
the Sidekiq (queues) Redis instance, which receives many BRPOP commands
with a timeout higher than the reporting threshold for this error.

As a result, when we enabled this flag we got flooded with 'expected'
errors for the queues Redis, which we didn't care about at all. So now
we also check which Redis instance we're using before reporting the
error.