Add starboard_vulnerability endpoint
These changes add an internal API endpoint for creating vulnerabilities
from Starboard. This will be used by the GitLab Kubernetes Agent in
order to track vulnerabilities found in running containers.

Since the service needed to do this is very similar to
ManuallyCreateService, I've also introduced a new subclass
(CreateServiceBase) which is now used as the base for
ManuallyCreateService and StarboardVulnerabilityCreateService.