Fix labels applied to a wrong issue
When "selectedLabels" prop changes,
the watcher for it in labels_select_root.vue
is supposed to update the vuex state "labels".

However, when the sidebar is hidden,
labels_select_root.vue component is destroyed
and the watcher is never used.

Note that we are in the process of deprecating
this particular implemenation of labels dropdown.

Changelog: fixed