Address vue i18n literal string violations
This is a follow-up to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/69622.

In dismiss_button.vue, the __() call was removed, since the argument
is already in a translated form via s__().

In modal_footer.vue, the __() call was removed, since the argument
is already in a translated form via s__().

In minutes_usage_project_chart.vue, the GlSprintf component was removed,
since it was effectively a no-op as it had no slots provided. The actual
rule violation has not been fixed, as it requires a bit more work to
solve correctly.