Replace vsa stage slug with id
Previously, default value stream stages did not have an `id` when
they were initialized, the `id` is assigned after a change
is made to the value stream and it is persisted. To work around
this we would use the value stream `title` instead, but this
causes issues when the user's language is changed.

Default value stream stages now have an `id` so we should
always use that instead.

Changelog: fixed
EE: true