Fix stale MR discussion cache when author role changes
Currently, when MR discussion is cached and the author's role has
changed (e.g. Developer to Guest), the role pill and some actions
will show as enabled (e.g. being able to resolve a discussion
but it'll not actually work).

This fix adds the max access of author to the cache key so when it
changes, the cache will be invalidated too.