Remove User#triggers dependent: :destroy
This does not appear to be necessary. It was first added in
https://gitlab.com/gitlab-org/gitlab-foss/-/merge_requests/9713#note_24835526
and there does not appear to be any justification for this as there is
already a cascading foreign key delete. There don't appear to be any
`before_destroy` or similar steps that might benefit from this
`dependent: :destroy` so it seems safe to remove it.