Set different session cookie for Geo secondaries
To support the secondary proxying the primary while also serving
and doing authentication on the secondary itself, the sites need
to have different session cookies in order to not overwrite each other.

As the primary site and the secondary site don't share the same
session state, when the cookie used on one site, is used with
the exact same name on the other site, it overwrites the content,
effectively logging out each other.

Changelog: changed
EE: true