Enable Sidekiq job compression without limiting
This allows compressing jobs in compression mode, without having to
specify a threshold at which to raise an error.

With this, setting a the limit threshold to 0 in compression mode,
will result in the jobs being compressed without anything else.

As a result of this, we will be able to enable the compression mode by
default for self-managed installations, who can all benefit from
this. But without having to enforce the same limits we do on GitLab.com