Remove File-By-File preference cookie
Since we now use a database setting to store the user's preference,
the cookie is redundant and can lead to some minor broken edge-cases.

Removing this cookie can also lead to a small annoyance in some
edge cases (see link below), but it's easily - and permanently -
remediated by toggling the checkbox again (or changing the user
preference in the settings page directly).

https://gitlab.com/gitlab-org/gitlab/-/issues/327162#note_549832248

Changelog: changed