First attempt at testing for descriptors
Attempt to start testing for a few items we use to describe
elements on a page, and recommend the correct words or phrases
instead. I changed the format of the returned message to give us more
flexibility in describing how to rewrite the sentence. Not all of
them are straightforward swaps.