Fix DB load balancing support for ActionCable
ActionCable workers do not run in the context of a request so we have to
add these callbacks that mimic our Rack / Sidekiq middleware.

We force the use of the primary because we currently have no way to
check the WAL location of the event that triggered the subscription and
these subscriptions are triggered in response to a write so up-to-date
information is important.