Fetch Ultimate plan ID from CustomersDot for trial prompts
- `TrialStatusWidgetHelper#ultimate_plan_id`
- `PaidFeatureCalloutHelper#premium_plan_id`
- Use the `GitlabSubscription::FetchSubscriptionPlansService` for each
  of those cases

Note: The `PaidFeatureCalloutHelper` is part of an experiment which is
currently unreachable
(https://gitlab.com/gitlab-org/gitlab/-/issues/273627#note_674150108)
but is yet to be cleaned up. It is likely that the code for it will be
entirely removed. If not, there is already an issue in place to extract
the commonalities between these two helpers
(https://gitlab.com/gitlab-org/gitlab/-/issues/340441), including
reusing the `FetchSubscriptionPlansService` data:
https://gitlab.com/gitlab-org/gitlab/-/blob/8938a54040c52454f58562626bf38584462f75c0/ee/app/helpers/trial_upgrade_prompts_helper.rb#L35