Add CI/CD variables for Auto Build and Auto Deploy image versions
By making the versions CI/CD variables, they can be set at the instance
level, at any group level, at the project level, and at the pipeline
level.

The main motivation is faster rollbacks of changes to Auto Build and
Auto Deploy on GitLab.com in the case of an incident, see
https://gitlab.com/gitlab-org/gitlab/-/issues/339033.

For example, to pin the version of auto-deploy-image to v2.12.0 for an
entire GitLab instance, run:

   curl --request POST --header "PRIVATE-TOKEN: <your_access_token>" \
     "https://gitlab.example.com/api/v4/admin/ci/variables" \
     --form "key=AUTO_DEPLOY_IMAGE_VERSION" \
     --form "value=v2.12.0"

Group level variables, and project level variables, can be set in their
respective CI/CD section of their settings.

Changelog: changed