Docs: Remove mention of rails env changing logs
The different naming of production{,_json}.log depending on the rails
environment is not important or useful for end users, which this
documentation is aimed at.

Additionally, it was not well explained in these docs before. We decided
that it was better to remove these lines than to explain them better.