tags: Always enable fix for verification of long tag messages
In contrast to signatures for commits, the signature of a tag is
directly attached to the tag message. So to get the tag signature,
GitLab loads the tag mesasge from Gitaly and extracts the signature from
it. However, if the tag message exceeds a certain length, Gitaly will
truncate it and thus keep us from verifying the signature.

This was fixed via a new `GetTagSignatures()` RPC in Gitaly, whose use
is currently guarded by the `:get_tag_signatures` feature flag. This new
implementation was rolled out to production on September 7th without any
observed issues and can thus be considered stable.

Remove the feature flag to always use the fix.

Changelog: fixed