Switch to Vue without runtime compiler
This allows us to ship a much smaller build of Vue in production and
forces folks to use Vue SFC with templates rather than defining
templates in HAML which is prone to XSS due to an additional layer of
indirection. We are already linting against new usage of this pattern
with our eslint rule: `@gitlab/no-runtime-template-compiler`

We still support compilation from Strings in Jest tests though.