Re-organize Environment Action Buttons
As part of the ongoing revitalization of environment management, some
re-organization and clarification of the action buttons for environments
is needed.

Many non-well used actions have been moved to a kebab menu. For the time
being, the "Deploy to" dropdown is still external. Pending design work
from the Pajamas team on a nested dropdown or menu component, it will be
moved to extra actions in the future.

Changelog: changed