Handle ConnectionNotEstablished in the DB LB
The database load balancer would re-raise ConnectionNotEstablished
errors when encountering them. This can lead to user facing errors
whenever the application tries to connect to a new database that is not
(yet) online.

This was discovered while testing the changes for
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/68042, after
adding a replica host that didn't exist.

Changelog: fixed