Implement rate-limiting for a deprecated API endpoint
GET /api/v4/groups/:id?with_projects=true is deprecated, as it includes
the `projects` member in the returned API response. That's expensive to
calculate and is generally unnecessary - other endpoints let you access
the same information in a more efficient manner.

Limiting requests to deprecated API endpoints is a way to induce users
to switch to the non-deprecated alternatives. We can add more endpoints
over time.

Changelog: added