Link to a search for feature flag name in project
We can utilize the power of search to find references to the feature
flag name in the code base to help make cleaning up the feature flag
easier.

The simple search for the name can be expanded on later by checking if
advanced search is enabled and utilizing some of the powers of elastic
search to refine our searching of the codebase.

Changelog: added
EE: true