Add the idea of a regular throttle
Currently we have three throttles that are exactly copy-pastes of each
other. I'm about to add a fourth, and the web throttle can become a
fifth once a couple of columns in the application_settings table are
renamed.

To avoid further copy-pasting in what is becoming a popular area of the
codebase, add a "regular throttle" abstraction that lets us just add
the throttle fragment / name to an array and get this functionality for
free.

Irregular throttles can still be added, but if we can make a throttle
regular, we should, to aid us in future refactorings.