Use the correct project path in generated `KUBECONFIG` file
Previously an Agent had to be directly associated with a project
for it to be used for the CI tunnel, which meant that the Agent
configuration project was the same project that was running the build.

With the introduction of authorizations for arbitrary groups and
projects, this assumption no longer holds. We must instead use the
project associated with the Agent instead of the one associated with
the build.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/70452

Changelog: fixed
EE: true