Measure image scaler duration for cached images
With this change, we measure durations in Prometheus
(in `imageResizeDurations`) whenever the requested image has not been modified and hence the cached version can be used by the client.

In this case, the image resizer process is not invoked, but
the client gets to see the image anyway.

As apdex is a measure of user satisfaction, it should not matter “where does the image come from?”, rather it should only care about “how soon is the image served”, which is why we have now decided toinclude the measurement of “cached versions” delivery in this apdex.

Changelog: changed