Make Redis instrumentation spec more robust
When run as the first spec in a run, this would fail. It would then pass
on retry. That was because on the first Redis call, we SELECT the
correct database for the test environment.

To work around this, we perform a Redis command - also doing the SELECT
if needed - and then reset the request store data so that the expected
counts can remain exact.