Replace "Any branch" with "All branches"
To improve clarity in the context of MR approval rules, this replaces
the phrase "Any branch" with the phrase "All branches". Now the
language is slightly more straightforward, with phrasing like
"Apply this approval rule to all branches" instead of
"Apply this approval rule to any branch" (which could be confusing).

Changelog: changed