Test project page load with unrecognized LICENSE
We've had an incident [1] where the project page gave 500 in case Gitaly
returned a license the Licensee gem did not recognize. This issue was
fixed with [2].

To prevent this error to happen again, add another test to verify if the
project page still loads when license detected by Gitaly is not
recognized by Licensee.

1. https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5478
2. https://gitlab.com/gitlab-org/gitlab/-/merge_requests/69878

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/340288