gitaly_client: Drop use of squash ID
In the past, the `UserSquash` RPC used to create worktrees in order to
compute the squashed commit. Starting with Gitaly v14.2.0, the RPC was
converted to never create worktrees anymore but instead to always
compute the commit in-memory. The squash ID, which used to indicate to
Gitaly how the worktree should be named and was required such that
`IsSquashInProgress()` knew which worktree to target, is thus not used
at all anymore and callers of `IsSquashInProgress()` have since been
removed.

Drop support for the deprecated squash ID such that it can eventually be
removed from Gitaly's protobuf definitions.