Add cloak-startup class and apply to mobile overlay
We currently have several specific elements cloaked to prevent
flashing when using startup.css. This MR introduces a generic
class which can be applied to any element to achieve the same.

Here we apply it to the mobile top nav overlay to fix broken
rendering in certain race/network conditions.

Changelog: fixed