Remove expired agent activity events as new events are created
Whenever an activity event is created for an agent, a worker is
scheduled to remove all but the most recent 200 events.

If the agent does not yet have 200 events, the worker still runs,
however it will not delete any records.

Events may be created frequently, so to prevent many workers running
unnecessarily they are scheduled with a delay of up to an hour, and
then de-duplicated until executed.

Changelog: added