Use rate limiting Redis for ApplicationRateLimiter
Here we can and do use a feature flag. The maximum interval for any rate
limit in this section is 3 minutes, so we don't mind too much about
requests that sneak in the gaps when the feature flag is being changed -
especially as feature flags are cached for a minute in process memory
anyway, so it would be hard to be much more precise.