Remove cross-joins from Project#upstream_projects,#downstream_projects
These methods were joining between `ci_*` and non `ci_*` tables which is
not going to be allowed when we move `ci_*` tables to a new database.
`disable_joins` was chosen here as this does not increase the
cardinality of data being returned as the only usage of this (aside
from counts handled in the next commit) is
[`project.downstream_projects.each`](
https://gitlab.com/gitlab-org/gitlab/-/blob/0d75b42c4ce663e4a9095348f97e6b84b55ee075/ee/app/services/ci/trigger_downstream_subscription_service.rb#L6
). This is fine to switch to `disable_joins` as there is no pagination,
no limits and no filtering. You can read more at
https://docs.gitlab.com/ee/development/database/multiple_databases.html#use-disable_joins-for-has_one-or-has_many-through-relations
.