Replace namespaces unique index on name and parent_id
This replaces the unique index on name and parent_id in
namespaces table with an unique index on name, parent_id
and type. We need this index change in order to cover name
uniqueness across namespaces of type Group and also allow
the possibility to have one namespace of type ProjectNamespace
with the same name as a Group to maintain current functionality
parity between namespaces and projects.

https://gitlab.com/gitlab-org/gitlab/-/issues/341814

Changelog: changed