Move the refresh header query to parent
As part of https://gitlab.com/gitlab-org/gitlab/-/issues/228746 we
are changing the fetching of a vulnerability to GraphQL. Currently,
it's being injected through HAML. Since the GraphQL fields are still
not complete, we decided to go step by step and work on this iteratively.
The first step is moving the refetching of the header to the vulnerability
component. UI-wise nothing significant changes here.