Don't run WAL queries when not using replicas
This changes the load balancer to not run redundant WAL related queries
when no replicas are in use. These queries are not needed as a single
database is always in sync with itself. It should also ensure the new
load balancer setup works on Aurora, at least for the time being.

Changelog: fixed