Order Deployments by Finish Time
If manual deployments are deployed out of order, the ordering of the
table gets murky and it is hard to tell which deployment was most
recently deployed.

Instead we can update the deployment ordering so that deployments are
ordered by when they finished at, which forces the finished deployments
to be sorted by most recently finished (read: deployed). Postgres ORDER
BY pushes NULL to the top of the list, so created/running (upcoming)
deployments are first

Changelog: changed