Fix inline JavaScript HAML linter not working
In https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/28570, we
moved the `haml_lint` directory out of an auto-load path to avoid a
startup failure in production
(https://gitlab.com/gitlab-org/gitlab-foss/-/issues/62125). However,
`inline_javascript.rb` used `require_dependency`, which is meant for
auto-loading and should not be mixed with `require`. As a result, the
linter registry was being reset twice: once for the `require_dependency`
call and once for the `require` call. This effectively dropped the
inline JavaScript linter.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/342479

Changelog: fixed