Gracefully track errors raised by sending CI minutes notifications
Sometimes while calculating the notification level to be sent to
the user we may see some errors. For example related to a project
being delete in the meantime.

We are rescuing and track such errors since we need to guarantee
that the service which updates CI minutes remains idempotent.

Changelog: fixed
EE: true